+++ app/Http/Livewire/Users.php
@@ -4,6 +4,8 @@ namespace App\Http\Livewire;
 
 use App\Models\User;
 use Exception;
+use Illuminate\Http\Client\Response;
+use Illuminate\Support\Facades\Gate;
 use Livewire\Component;
 use Livewire\WithPagination;
 
@@ -12,8 +14,17 @@ class Users extends Component
     use WithPagination;
     public $selected=[];
     public $deleted_record;
+    public function mount()
+    {
+        if(Gate::denies('user_access')){
+            redirect()->route('dashboard');
+        }
+    }
     public function Delete($id)
     {
+        if(Gate::denies('user_delete')){
+            redirect()->route('dashboard');
+        }
         $user = User::find($id);
         $this->deleted_record = $user->name;
         $user->delete();
@@ -21,6 +32,9 @@ class Users extends Component
     }
     public function DeleteSelected()
     {
+        if(Gate::denies('user_delete')){
+            redirect()->route('dashboard');
+        }
         User::destroy($this->selected);
         $this->deleted_record = "Selected Users";
         $this->emitSelf('record_deleted');
