+++ resources/views/livewire/portfolio/all-messages.blade.php
@@ -1,4 +1,5 @@
 <div class="">
+    @can('inbox_access')
     <x-table.action_performed event-name="message_deleted" :details="$deleted_message"></x-table.action_performed>
     <div class="grid md:grid-cols-10 mt-5 mb-2 min-h-[570px]">
         <div class="hidden md:block col-span-3 h-full w-full pr-3">
@@ -26,11 +27,13 @@
                             <div class="flex justify-between">
                                 <span wire:click="MessageToShow({{$message->id}})" class="text-sm font-semibold text-gray-900 dark:text-gray-200">{{$message->name}}</span>
                                 <div class="flex space-x-2 text-sm text-gray-400 dark:text-gray-500">
-                                    <span wire:click="DeleteMessage({{$message->id}})" class="hover:text-gray-600 dark:hover:text-gray-400">
-                                        <svg xmlns="http://www.w3.org/2000/svg" class="h-5 w-5" fill="none" viewBox="0 0 24 24" stroke="currentColor">
-                                             <path stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M19 7l-.867 12.142A2 2 0 0116.138 21H7.862a2 2 0 01-1.995-1.858L5 7m5 4v6m4-6v6m1-10V4a1 1 0 00-1-1h-4a1 1 0 00-1 1v3M4 7h16" />
-                                        </svg>
-                                    </span>
+                                    @can('message_delete')
+                                        <span wire:click="DeleteMessage({{$message->id}})" class="hover:text-gray-600 dark:hover:text-gray-400">
+                                            <svg xmlns="http://www.w3.org/2000/svg" class="h-5 w-5" fill="none" viewBox="0 0 24 24" stroke="currentColor">
+                                                <path stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M19 7l-.867 12.142A2 2 0 0116.138 21H7.862a2 2 0 01-1.995-1.858L5 7m5 4v6m4-6v6m1-10V4a1 1 0 00-1-1h-4a1 1 0 00-1 1v3M4 7h16" />
+                                            </svg>
+                                        </span>
+                                    @endcan
                                     <span wire:click="ChangeMark({{$message->id}})" title="Read Message" class="text-sm text-gray-400 hover:text-gray-600">
                                         <svg xmlns="http://www.w3.org/2000/svg" class="h-5 w-5" fill="none" viewBox="0 0 24 24" stroke="currentColor">
                                             <path stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M3 19v-8.93a2 2 0 01.89-1.664l7-4.666a2 2 0 012.22 0l7 4.666A2 2 0 0121 10.07V19M3 19a2 2 0 002 2h14a2 2 0 002-2M3 19l6.75-4.5M21 19l-6.75-4.5M3 10l6.75 4.5M21 10l-6.75 4.5m0 0l-1.14.76a2 2 0 01-2.22 0l-1.14-.76" />
@@ -50,11 +53,13 @@
                             <div class="flex justify-between">
                             <span wire:click="MessageToShow({{$message->id}})" class="text-sm font-semibold text-gray-900 dark:text-gray-200">{{$message->name}}</span>
                             <div class="flex space-x-2 text-sm text-gray-400 dark:text-gray-500">
+                                @can('message_delete')
                                 <span wire:click="DeleteMessage({{$message->id}})" class="hover:text-gray-600 dark:hover:text-gray-400">
                                     <svg xmlns="http://www.w3.org/2000/svg" class="h-5 w-5" fill="none" viewBox="0 0 24 24" stroke="currentColor">
                                         <path stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M19 7l-.867 12.142A2 2 0 0116.138 21H7.862a2 2 0 01-1.995-1.858L5 7m5 4v6m4-6v6m1-10V4a1 1 0 00-1-1h-4a1 1 0 00-1 1v3M4 7h16" />
                                     </svg>
                                 </span>
+                                @endcan
                                 <span wire:click="ChangeMark({{$message->id}})" class="hover:text-gray-600 dark:hover:text-gray-400">
                                     <svg xmlns="http://www.w3.org/2000/svg" class="h-5 w-5" fill="none" viewBox="0 0 24 24" stroke="currentColor">
                                         <path stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M3 8l7.89 5.26a2 2 0 002.22 0L21 8M5 19h14a2 2 0 002-2V7a2 2 0 00-2-2H5a2 2 0 00-2 2v10a2 2 0 002 2z" />
@@ -119,4 +124,5 @@
         </div>
     </div>
     {{ $messages->links('livewire.admin.admin-pagination') }}
+    @endcan
 </div>
